Replace standard json package with github.com/goccy/go-json
Note:
go-json has some bugs that have been fixed in github.com/peterlimg/go-json. We will use the code on peterlimg/go-json by using the replace command on go.mod file. Will set it back to goccy/go-json when they merge the PR for the fix or they find the way themselves to fix it.